todo controller works

+++ ToDoApp Project/View/UsersManagementView.cs
@@ -31,10 +31,30 @@ namespace ToDoApp_Project.View
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            DataGridViewRow row = dgvUserManagementView.CurrentRow;
-            int id = int.Parse(row.Cells[0].Value.ToString());
-            userController.DeleteUser(id);
-            RefreshTable();
+            DialogResult result = MessageBox.Show("Do really want to delete this user?", "IMPORTANT",
+                MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
+            if (result == DialogResult.OK)
+            {
+                DialogResult result2 = MessageBox.Show("By accepting this, you understand that every ToDo that this user created is going to be deleted too!", "IMPORTANT",
+                    MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
+                if (result2 == DialogResult.OK)
+                {
+                    DataGridViewRow row = dgvUserManagementView.CurrentRow;
+                    int id = int.Parse(row.Cells[0].Value.ToString());
+                    userController.DeleteUser(id);
+                    RefreshTable();
+                }
+                else if (result2 == DialogResult.Cancel)
+                {
+                    MessageBox.Show("No user was deleted.", "Response",
+                        MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+            }
+            else if (result == DialogResult.Cancel)
+            {
+                MessageBox.Show("No user was deleted.", "Response",
+                        MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
         }
 
         private void btnGoBack_Click(object sender, EventArgs e)
@@ -46,47 +66,48 @@ namespace ToDoApp_Project.View
 
         private void btnCreate_Click(object sender, EventArgs e)
         {
-            User user = new User();
-            string username = txtBoxUsername.Text;
-            string password = txtBoxPassword.Text;
-            string userrole = txtBoxRole.Text;
-            int id = int.Parse(txtBoxId.Text);
-
-            if (!string.IsNullOrEmpty(username) && !string.IsNullOrEmpty(password)
-                && !string.IsNullOrEmpty(userrole) && !string.IsNullOrEmpty(id.ToString()))
+            if (string.IsNullOrEmpty(txtBoxUsername.Text) || string.IsNullOrEmpty(txtBoxPassword.Text)
+                || string.IsNullOrEmpty(txtBoxRole.Text) || string.IsNullOrEmpty(txtBoxId.Text))
             {
-                user.Username = username;
-                user.Password = password;
-                user.Role = userrole;
-                user.Id = id;
-                if (userController.doesUsernameExist(user))
-                {
-                    MessageBox.Show("Username already exists!", "Username is being used",
+                MessageBox.Show("Please dont leave the text boxes empty!", "EMPTY BOX DETECTED",
                         MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                    txtBoxUsername.Text = "";
-                    txtBoxPassword.Text = "";
-                    txtBoxRole.Text = "";
-                    txtBoxId.Text = "";
-                }
-                else if (userController.doesIdExist(user))
-                {
-                    MessageBox.Show("Id is being used!", "Id error",
-                        MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                    txtBoxUsername.Text = "";
-                    txtBoxPassword.Text = "";
-                    txtBoxRole.Text = "";
-                    txtBoxId.Text = "";
-                }
-                else
-                {
-                    userController.RegisterUser(user);
-                    RefreshTable();
-                }
+                txtBoxUsername.Text = "";
+                txtBoxPassword.Text = "";
+                txtBoxRole.Text = "";
+                txtBoxId.Text = "";
+            }
+            else if (userController.doesUsernameExist(txtBoxUsername.Text))
+            {
+                MessageBox.Show("Username already exists!", "PROVIDED DATA ERROR",
+                        MessageBoxButtons.OK, MessageBoxIcon.Error);
+                txtBoxUsername.Text = "";
+                txtBoxPassword.Text = "";
+                txtBoxRole.Text = "";
+                txtBoxId.Text = "";
+            }
+            else if (userController.doesIdExist(int.Parse(txtBoxId.Text)))
+            {
+                MessageBox.Show("Id is being used!", "PROVIDED DATA ERROR",
+                        MessageBoxButtons.OK, MessageBoxIcon.Error);
+                txtBoxUsername.Text = "";
+                txtBoxPassword.Text = "";
+                txtBoxRole.Text = "";
+                txtBoxId.Text = "";
             }
             else
             {
-                MessageBox.Show("Please dont leave the text boxes empty!", "EMPTY BOX DETECTED",
-                        MessageBoxButtons.OK, MessageBoxIcon.Information);
+                User user = new User();
+                user.Username = txtBoxUsername.Text;
+                user.Password = txtBoxPassword.Text;
+                user.Role = txtBoxRole.Text;
+                user.Id = int.Parse(txtBoxId.Text);
+
+                userController.RegisterUser(user);
+                txtBoxUsername.Text = "";
+                txtBoxPassword.Text = "";
+                txtBoxRole.Text = "";
+                txtBoxId.Text = "";
+                RefreshTable();
             }
         }
 
@@ -95,51 +116,79 @@ namespace ToDoApp_Project.View
             DataGridViewRow row = dgvUserManagementView.CurrentRow;
             int id = int.Parse(row.Cells[0].Value.ToString());
 
-            User user = new User();
-            string username = txtEditName.Text;
-            string password = txtEditPass.Text;
-            string role = txtEditRole.Text;
-
-            if (!string.IsNullOrEmpty(username) && !string.IsNullOrEmpty(password)
-                && !string.IsNullOrEmpty(role))
+            if (string.IsNullOrEmpty(txtEditName.Text) || string.IsNullOrEmpty(txtEditPass.Text)
+                || string.IsNullOrEmpty(txtEditRole.Text))
             {
-                user.Username = username;
-                user.Password = password;
-                user.Role = role;
-
-                if (userController.doesUsernameExist(user))
-                {
-                    MessageBox.Show("Username already exists!", "Username is being used",
+                MessageBox.Show("Please dont leave the text boxes empty!", "EMPTY BOX DETECTED",
                         MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                    txtBoxUsername.Text = "";
-                    txtBoxPassword.Text = "";
-                    txtBoxRole.Text = "";
-                    txtBoxId.Text = "";
-                }
-                else
-                {
-                    userController.EditUser(id, user);
-                    RefreshTable();
-                }
+                txtEditName.Text = "";
+                txtEditPass.Text = "";
+                txtEditRole.Text = "";
+            }
+            else if(userController.doesUsernameExist(txtEditName.Text))
+            {
+                MessageBox.Show("Username already exists!", "PROVIDED DATA ERROR",
+                        MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                txtEditName.Text = "";
+                txtEditPass.Text = "";
+                txtEditRole.Text = "";
             }
             else
             {
-                MessageBox.Show("Please dont leave the text boxes empty!", "EMPTY BOX DETECTED",
-                        MessageBoxButtons.OK, MessageBoxIcon.Information);
+                User user = new User();
+                user.Username = txtEditName.Text;
+                user.Password = txtEditPass.Text;
+                user.Role = txtEditRole.Text;
+
+                userController.EditUser(id, user);
+                txtEditName.Text = "";
+                txtEditPass.Text = "";
+                txtEditRole.Text = "";
+                RefreshTable();
             }
         }
 
         private void btnDeleteById_Click(object sender, EventArgs e)
         {
-            int id = int.Parse(txtBoxDeleteById.Text);
-            userController.DeleteUser(id);
-            txtBoxDeleteById.Text = "";
-            RefreshTable();
-        }
-
-        private void lblDeleteManagement_Click(object sender, EventArgs e)
-        {
-
+            if (string.IsNullOrEmpty(txtBoxDeleteById.Text) || string.IsNullOrWhiteSpace(txtBoxDeleteById.Text))
+            {
+                MessageBox.Show("Please dont leave the text boxes empty!", "EMPTY BOX DETECTED",
+                        MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
+            else if (!userController.doesIdExist(int.Parse(txtBoxDeleteById.Text)))
+            {
+                MessageBox.Show($"User with Id: ${txtBoxDeleteById.Text} doesn't exist!", "PROVIDED DATA ERROR",
+                        MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else
+            {
+                DialogResult result = MessageBox.Show("Do really want to delete this user?", "IMPORTANT",
+                    MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
+                if (result == DialogResult.OK)
+                {
+                    DialogResult result2 = MessageBox.Show("By accepting this, you understand that every ToDo that this user created is going to be deleted too!", "IMPORTANT",
+                        MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
+                    if (result2 == DialogResult.OK)
+                    {
+                        int id = int.Parse(txtBoxDeleteById.Text);
+                        userController.DeleteUser(id);
+                        txtBoxDeleteById.Text = "";
+                        RefreshTable();
+                    }
+                    else if (result2 == DialogResult.Cancel)
+                    {
+                        MessageBox.Show("No user was deleted.", "Response",
+                            MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        txtBoxDeleteById.Text = "";
+                    }
+                }
+                else if (result == DialogResult.Cancel)
+                {
+                    MessageBox.Show("No user was deleted.", "Response",
+                            MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    txtBoxDeleteById.Text = "";
+                }
+            }
         }
     }
 }
